+++ src/p_client.c
@@ -488,7 +488,7 @@ void LookAtKiller(edict_t *self, edict_t *inflictor, edict_t *attacker)
 
     // PMM - fixed to correct for pitch of 0
     if (dir[0])
-        self->client->killer_yaw = RAD2DEG(atan2(dir[1], dir[0]));
+        self->client->killer_yaw = RAD2DEG(atan2f(dir[1], dir[0]));
     else if (dir[1] > 0)
         self->client->killer_yaw = 90;
     else if (dir[1] < 0)
